+++ packages/launchql-gen/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.10](https://github.com/launchql/launchql/compare/@launchql/codegen@2.5.9...@launchql/codegen@2.5.10) (2025-05-27)
+
+**Note:** Version bump only for package @launchql/codegen
+
+
+
+
+
 ## [2.5.9](https://github.com/launchql/launchql/compare/@launchql/codegen@2.5.8...@launchql/codegen@2.5.9) (2025-05-25)
 
 **Note:** Version bump only for package @launchql/codegen