+++ packages/pg-codegen/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.5](https://github.com/launchql/launchql/compare/pg-codegen@2.2.4...pg-codegen@2.2.5) (2025-05-27)
+
+**Note:** Version bump only for package pg-codegen
+
+
+
+
+
 ## [2.2.4](https://github.com/launchql/launchql/compare/pg-codegen@2.2.3...pg-codegen@2.2.4) (2025-05-25)
 
 **Note:** Version bump only for package pg-codegen